OneShotHaiku/requirements.txt
python-chess==1.10.0
